+++ packages/moonstone/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 The following is a curated list of changes in the Enact moonstone module, newest changes on the top.
 
+## [unreleased]
+
+### Changed
+
+- `moonstone/ExpandableList` and `moonstone/Repeater` to accept an array of objects as children, which are spread onto the generated components
+
 ## [1.0.0-beta.4] - 2017-03-10
 
 ### Added